api-3/requirements.txt
Flask
Flask-Cors
pandas
openai
python-dotenv
mysql-connector-python
